Reformatting and PSR changes

+++ back-end/index.php
@@ -13,100 +13,59 @@
 
   // Tries to get APIkey
   try {
-
+    // Throws error if header not available
     $headers = apache_request_headers();
-    $dbConn->checkAPIkey($headers['APIKEY']);
-
+    if(!array_key_exists('APIKEY', $headers)){
+      throw new \Exception("No API key in header.");
+    }
+    $apiKey = $headers['APIKEY'];
+    $dbConn->checkAPIkey($apiKey);
   } catch (Exception $e) {
-
-    $output['status']['code'] = "401";
-    $output['status']['name'] = "ok";
-    $output['status']['description'] = "Auth failed.";
-    $output['data'] = "No API key in header.";
-
-    $dbConn->conn = NULL;
-
-    echo json_encode($output);
-
-    exit;
+    $dbConn->outputError("401", "error", "Auth failed.", $e->getMessage());
   }
 
   // Server uses HTTP verbs to determine actions (GET, POST, PUT, DELETE)
   if ($_SERVER['REQUEST_METHOD'] === 'GET') {
-
     $url = parse_url($_SERVER['REQUEST_URI']);
     if(array_key_exists('id', $_GET)) {
-
       $id = $_GET['id'];
       $dbConn->getUser($id);
-
     } else if (!array_key_exists('query', $url)) {
-
       $dbConn->getUsers();
-
     }
-
     else {
-
-      outputURLError($dbConn);
-
+      $dbConn->outputURLError();
     }
-
   }
 
   // HTTP verb for inserting users
   if ($_SERVER['REQUEST_METHOD'] === 'PUT') {
-
     $dbConn->insertUser();
-
   }
 
   // HTTP verb for updating users
   if ($_SERVER['REQUEST_METHOD'] === 'POST') {
-
     if(array_key_exists('id', $_POST)) {
-
       $id = $_POST['id'];
       $dbConn->updateUser($id);
-
     } else {
-
-      outputURLError($dbConn);
-
+      $dbConn->outputURLError();
     }
 
   }
 
   // HTTP verb for deleting users
   if ($_SERVER['REQUEST_METHOD'] === 'DELETE') {
-
     try {
-
       $url = parse_url($_SERVER['REQUEST_URI']);
+      if(!array_key_exists('query', $url)){
+        throw new \Exception("No query string for id.");
+      }
       parse_str($url['query'], $DELETE);
       $id = $DELETE['id'];
       $dbConn->deleteUser($id);
-
     } catch (Exception $e) {
-
-      outputURLError();
-
+      $dbConn->outputURLError();
     }
-
-  }
-
-  // Returns error codes, closes dbConn and exits script for URL error
-  function outputURLError($dbConn)
-  {
-    $output['status']['code'] = "400";
-    $output['status']['name'] = "executed";
-    $output['status']['description'] = "query failed";
-    $output['data'] = "URL string has incorrect format.";
-
-    echo json_encode($output);
-
-    $dbConn->conn = NULL;
-
-    exit;
 
   }
